+++ Src/GuiStracini.SDKBuilder/BaseRequest.cs
@@ -11,10 +11,10 @@
 // </copyright>
 // <summary></summary>
 // ***********************************************************************
-namespace GuiStracini.SDKBuilder
-{
-    /// <summary>
-    /// The base request class
-    /// </summary>
-    public abstract class BaseRequest { }
-}
+
+namespace GuiStracini.SDKBuilder;
+
+/// <summary>
+/// The base request class
+/// </summary>
+public abstract class BaseRequest { }
