+++ Src/GuiStracini.SDKBuilder/BaseResponse.cs
@@ -11,23 +11,23 @@
 // </copyright>
 // <summary></summary>
 // ***********************************************************************
-namespace GuiStracini.SDKBuilder
+
+namespace GuiStracini.SDKBuilder;
+
+/// <summary>
+/// The base response class.
+/// </summary>
+public abstract class BaseResponse
 {
     /// <summary>
-    /// The base response class.
+    /// Gets or sets the error code.
     /// </summary>
-    public abstract class BaseResponse
-    {
-        /// <summary>
-        /// Gets or sets the error code.
-        /// </summary>
-        /// <value>The error code.</value>
-        public int ErrorCode { get; set; }
+    /// <value>The error code.</value>
+    public int ErrorCode { get; set; }
 
-        /// <summary>
-        /// Gets or sets the error message.
-        /// </summary>
-        /// <value>The error message.</value>
-        public string ErrorMessage { get; set; }
-    }
+    /// <summary>
+    /// Gets or sets the error message.
+    /// </summary>
+    /// <value>The error message.</value>
+    public string ErrorMessage { get; set; }
 }
