+++ Src/GuiStracini.SDKBuilder/GuiStracini.SDKBuilder.csproj
@@ -2,11 +2,11 @@
 
   <PropertyGroup>
     <ProjectGuid>528839BC-251E-40CB-8FD3-79C7150DC060</ProjectGuid>
-    <TargetFrameworks>netstandard2.0;netstandard2.1</TargetFrameworks>
+    <TargetFrameworks>netstandard2.0;netstandard2.1;net8.0;net9.0</TargetFrameworks>
     <LangVersion>latest</LangVersion>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <Authors>Guilherme Branco Stracini</Authors>
-    <Copyright>© 2019 - 2023 Guilherme Branco Stracini. All rights reserved.</Copyright>
+    <Copyright>© 2019 - 2025 Guilherme Branco Stracini. All rights reserved.</Copyright>
     <Description>A SDK framework that helps create a C# SDK project for consuming a generic REST API based on Microsoft WebAPI 2 conventions</Description>    
     <PackageProjectUrl>https://guibranco.github.io/GuiStracini.SDKBuilder/</PackageProjectUrl>
     <RepositoryUrl>https://github.com/guibranco/GuiStracini.SDKBuilder</RepositoryUrl>
